chore: better fetch currency, update useTimelineItems hook
