+++ packages/localization/src/config/translations.json
@@ -3757,6 +3757,7 @@
   "Partial Success": "Partial Success",
   "Progress": "Progress",
   "Swapped %currencyA% to %currencyB% (%chainName%)": "Swapped %currencyA% to %currencyB% (%chainName%)",
+  "Swap %currencyA% to %currencyB% (%chainName%)": "Swap %currencyA% to %currencyB% (%chainName%)",
   "Bridge %currency% (%inputChain% to %outputChain%)": "Bridge %currency% (%inputChain% to %outputChain%)",
   "Partial Fee": "Partial Fee",
   "%amount% %symbol% has been sent to your wallet on %outputChain%": "%amount% %symbol% has been sent to your wallet on %outputChain%",
